requirements.txt
django==3.1.4
gunicorn==20.0.4
psycopg2==2.8.6
python_dotenv==0.15.0
djangorestframework==3.12.2
